part of debug

+++ data/augmentation.py
@@ -1,5 +1,3 @@
-import skimage.io as imgio
-from skimage.transform import resize
 import random
 import collections.abc as cabc
 import numpy as np
@@ -25,7 +23,7 @@ def randRescaleAndTranspose(scale, *originalArray):
             reshape[1] = shape[1] * shortLen / shape[0]
             reshape[0] = shortLen
         scaledImg = cv2.resize(img, (int(reshape[0]), int(reshape[1]))).transpose((2, 0, 1))
-        scaledImg=scaledImg.astype(np.float64)
+        #scaledImg=scaledImg.astype(np.float64)
         scaledImgs.append(scaledImg)
     return scaledImgs
